requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
motor>=3.3.0
pydantic>=2.6.0
pydantic-core>=2.16.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
